+++ module/Core/src/EventDispatcher/Event/VisitLocated.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Shlinkio\Shlink\Core\EventDispatcher\Event;
+
+final class VisitLocated extends AbstractVisitEvent
+{
+}